UCC-cleanup: Clean up the obvious
